message: replace LineIteratorExt::check_message_format with query fn
... returning a `bool`

The purpose of `check_message_format` was, as the name suggests, to
check whether the formatting of a message is correct. However,
because we wrote it with a very narrow use-case in mind, it returned a
`Result` which is arguably inappropriate considering broader use-cases.

This change rectifies the situation. It also removing the use of
`git2::Error` in a place that was awkward, considering planned future
changes towards decoupling from `git2`.