repository: make RepositoryExt generic over repo associated lifetime
Currently, all git-dit functionality is implemented based on `git2`. In
the (possibly far-ish) future, we also want to support `gix`. There are
obviously (sometimes big) differences between implementations when it
comes to the types they use for representing. We will therefore
introduce various associated types in the future.

For `git2`, those types will often carry a lifetime linking the entities
returned by some `RepositoryExt` fn to `&self` (or `&mut self`). We
could make the associated types themselves generic over some lifetime
and fill that parameter with the anonymous one when using it in some fn.
However, we we do not want to impose the same lifetime restrictions to
future additional impls (e.g. for `gix`). Instead, we just want to
specify the dependence for one specific impl if necessary.

We can do so, but for that we need a lifetime to work with. We therefore
introduce a lifetime as a parameter for `RepositoryExt`.