repository: add assiciated type for InnerError to RepositoryExt
Currently, all git-dit functionality is implemented based on `git2`. In
the (possibly far-ish) future, we also want to support `gix`. There are
obviously (sometimes big) differences between implementations when it
comes to the types they use for representing. We therefore decided to
introduce various associated types. One of the most basic ones is the
`Error` type returned by `git2::Repository` or whatever type we impl
`RepositoryExt` for.