test_utils: add fn for creating an empty tree
We currently follow a long term plan to decouple this library from
`git2` to a greater extend, with the goal to also support `gix` in the
future. The current `RepositoryExt::empty_tree` was introduced as a
convenience fn, because creating an empty tree is a bit tedious with
`git2`.

Aside from tests, we never actually create empty trees inside the
library. Instead, we expect any tree to be supplied by the (library)
user. For this reason, an agnostic `RepositoryExt` for _issue_
management does not need such a fn. We therefore decided to remove it
for now.

As for the tests, we still want a simple way for creating an empty tree.
This change introduces a fn for that purpose —which _may_ end up getting
integrated in another extension trait specifically for tests at some
point in the future.

Future changes will mograte tests away from `RepositoryExt::empty_tree`
until we can finally remove it alltogether.