Ignore test directory
We currently have tests that need to perform operations on a repository.
Historically, we'd put those repositories in a dedicated `test`
directory which is not part of your usual cargo setup. It used to be
ignored by the toplevel `.gitignore` when the library still lived in a
sub-directory rather then the repository's root. But that `.gitignore`
was removed in

	4719e86861bc687eb32965d61b07837682d3b1ee

When we yanked code for the git-dit executable from the tree.