traversal: add TraversalBuilder for building an Iterator over messages
We are currently using the `Messages` and other `Iterator`s based on
`git2::Revwalk` in quite a few places. Many of them live in the `iter`
submodule. However, we want to decouple git-dit from `git2` and support
other git implementations such as `gix` in the future. Our current
reliance on `git2::Revwalk` is obviously a roadblock that we need to get
rid of first.

Changing `Messages` and all its uses would be an option, but not one we
prefer. Rather, we decided to construct new utilities in parallel and
replace `Messages` bit by bit. We take the opportunity to move away from
an `Iterator` for commits towards `Iterator`s over object ids, as
looking up and retrieving the commit is often times not necessary and
just created additional overhead.

This change introduces a trait that we can use as an abstraction for
`Iterator` configuration. Because git implementations other than `git2`
may not allow configuration of a "revwalk" after creation, we make it a
builder abstraction.