repository: add fn for creating a traversal builder
We are currently using the `Messages` and other `Iterator`s based on
`git2::Revwalk` in quite a few places. Many of them live in the `iter`
submodule. However, we want to decouple git-dit from `git2` and support
other git implementations such as `gix` in the future. Our current
reliance on `git2::Revwalk` is obviously a roadblock that we need to get
rid of first.

We recently introduced a builder abtraction named `TraversalBuilder`.
This change now links that abstraction to the `RepositoryExt` via an
associated type and a function for creating a builder.