repository: add repo lifetime to all self params in RepositoryExt
Currently, all git-dit functionality is implemented based on `git2`. In
the (possibly far-ish) future, we also want to support `gix`. There are
obviously (sometimes big) differences between implementations when it
comes to the types they use for representing. We will therefore
introduce various associated types in the future.

For `git2`, those types will often carry a lifetime linking the entities
returned by some `RepositoryExt` fn to `&self` (or `&mut self`). For
this purpose, we recently introduced a lifetime parameter to the trait.
This change now introduces the parameter into the signatures of all
associated fns. We also applied style fixes to those signatures.