repository: use Self::InnerError in RepositoryExt fn signatures
Currently, all git-dit functionality is implemented based on `git2`. In
the (possibly far-ish) future, we also want to support `gix`. There are
obviously (sometimes big) differences between implementations when it
comes to the types they use for representing.

We therefore decided to introduce various associated types. We already
introduced a type for the `Error`s emitted by the underlying repository
implementation. This change now makes use of that type in fn signatures
for all relevant `RepositoryExt` fns.