error: rename InnerError::Reference to RefName
When representing references in `Error`s, we want something 'static and
"robust". In practice, we are more interested in the name of a reference
than anything else. Since the `git2::Reference` is neither 'static nor
does it implement `Display`, which is why we use `String`.

If we want to move more and more functionality into generic and/or
provided associated fns, we need to provide some mechanism to extract
the name of a reference. We plan to do this via the `InnerError` trait,
which makes the current name of `Reference` awkward.