repository: add associated type for References to RepositoryExt
Currently, all git-dit functionality is implemented based on `git2`. In
the (possibly far-ish) future, we also want to support `gix`. There are
obviously (sometimes big) differences between implementations when it
comes to the types they use for representing. We therefore decided to
introduce various associated types.

This change introduces a type for references.