repository: use Self::Rerefence in RepositoryExt fns
Currently, all git-dit functionality is implemented based on `git2`. In
the (possibly far-ish) future, we also want to support `gix`. There are
obviously (sometimes big) differences between implementations when it
comes to the types they use for representing.

We therefore decided to introduce various associated types. We already
introduced a type for (git) references as used by the underlying
implementation. This change now makes use of that type in fn signatures
for all relevant `RepositoryExt` fns.