repository: remove RepositoryExt::issue_messages_iter
We are currently using the `Messages` and other `Iterator`s based on
`git2::Revwalk` in quite a few places. Many of them live in the `iter`
submodule. However, we want to decouple git-dit from `git2` and support
other git implementations such as `gix` in the future. Our current
reliance on `git2::Revwalk` is obviously a roadblock that we need to get
rid of first.

Changing `Messages` and all its uses would be an option, but not one we
prefer. Rather, we decided to construct new utilities in parallel and
replace `Messages` (and things like `IssueMessagesIter`) bit by bit.
This change removes the one function in `RepositoryExt` for creating an
`IssueMessagesIter`, removing one block for changes to the signature of
`RepositoryExt::first_parent_messages`.