repository: make RepositoryExt::first_parent_messages a provided fn
We are currently using the `Messages` and other `Iterator`s based on
`git2::Revwalk` in quite a few places. Many of them live in the `iter`
submodule. However, we want to decouple git-dit from `git2` and support
other git implementations such as `gix` in the future. Our current
reliance on `git2::Revwalk` is obviously a roadblock that we need to get
rid of first.

Changing `Messages` and all its uses would be an option, but not one we
prefer. Rather, we decided to construct new utilities in parallel and
replace `Messages` bit by bit. This change makes `first_parent_messages`
a provided fn with an implementation based purely on the new
abstractions.