iter: remove Messages::until_any_initial
We are currently using the `Messages` and other `Iterator`s based on
`git2::Revwalk` in quite a few places. Many of them live in the `iter`
submodule. However, we want to decouple git-dit from `git2` and support
other git implementations such as `gix` in the future. Our current
reliance on `git2::Revwalk` is obviously a roadblock that we need to get
rid of first.

Currently, we try to eliminate `Messages` and `IssueMessagesIter`. This
change removes just `Messages::until_any_initial`, which allows us to
remove `IssueMessagesIter` in once change and `Messages` in another.