iter: remove IssueMessagesIter
We are currently using the `Messages` and other `Iterator`s based on
`git2::Revwalk` in quite a few places. Many of them live in the `iter`
submodule. However, we want to decouple git-dit from `git2` and support
other git implementations such as `gix` in the future. Our current
reliance on `git2::Revwalk` is obviously a roadblock that we need to get
rid of first.

Changing `Messages` and all its uses would have been an option, but
instead we constructed new utilities in parallel and migrated all users
of `Messages` and `IssueMessagesIter`. This change now removes the
latter, as its no longer used and too awkward to migrate to the new
`TraversalBuilder` abstraction.