base: add Base trait defining types of git impls
Currently, `libgitdit` is based on `git2`. We extend `git2::Prepository`
via our trait `RepositoryExt`. We do want to support using other git
implementations in the future, namely `gix` and are thus currently
introducing abstractions which allow removal of hard-coded `git2` types
from the trait.

We decided to take the opportunity to modularize a bit more and move
some functionality into different traits that are then implemented for
`git2::Repository`. The goal is for `RepositoryExt` to require those,
and for it to only have provided methods, i.e. only trivial (or empty)
impls.

Reconciling the various associated types of various different
super-traits is bound to get messy, given that there are quite a few
relations between them. We therefore decided to put at least the ones
most likely shared between traits in a separate `Base` trait.