repository: make RepositoryExt require Traversible
Currently, `libgitdit` is based on `git2`. We extend `git2::Prepository`
via our trait `RepositoryExt`. We do want to support using other git
implementations in the future, namely `gix` and are thus currently
introducing abstractions which allow removal of hard-coded `git2` types
from the trait.

We decided to take the opportunity to modularize a bit more and move
some functionality into different traits that are then implemented for
`git2::Repository`. The goal is for `RepositoryExt` to require those,
and for it to only have provided methods, i.e. only trivial (or empty)
impls.

We recently introduced the `Base` trait with a small set of associated
types and the `Traversible` trait with the `TraversalBuilder` type and a
fn for creating such an entity. This change now makes `RepositoryExt`
require those (via `Traversible`) and replaces the now redundant types
and fn.