traversal: add trait for Traversible entities
Currently, `libgitdit` is based on `git2`. We extend `git2::Prepository`
via our trait `RepositoryExt`. We do want to support using other git
implementations in the future, namely `gix` and are thus currently
introducing abstractions which allow removal of hard-coded `git2` types
from the trait.

We decided to take the opportunity to modularize a bit more and move
some functionality into different traits that are then implemented for
`git2::Repository`. The goal is for `RepositoryExt` to require those,
and for it to only have provided methods, i.e. only trivial (or empty)
impls.

We recently introduced an abstraction for traversing commits. Or rather,
an abstraction for building a traversing `Iterator`. `RepositoyExt` thus
gained an associated type and fn. This change now adds a trait with that
exact type and fn, including an impl for `git2::Repository`.

A future change will make `RepositoryExt` require this trait.