repository: move fn first_parent_messages to Traversible
We decided to modularize a bit more and move some functionality into
different traits that are then implemented for `git2::Repository`. The
goal is for `RepositoryExt` to require those, and for it to only have
provided methods, i.e. only trivial (or empty) impls. In our pursuit of
this goal, we recently introduced the `Traversible` trait with the fn
`traversal_builder` and made `RepositoryExt` require it.

While `first_parent_messages` is already a provided fn that makes use of
`traversal_builder`, it is primarily an internal conveinence fn, and it
does make use _only_ of `traversal_builder` and some properties of the
`InnerError`. We therefor decided that is really ought to be part of
`Traversible`.

This change moves the fn and, while we're at it, adjusts its
documentation to reflect the new reality.