repository: do not make RepositoryExt require Traversible
We decided to modularize a bit more and move some functionality into
different traits that are then implemented for `git2::Repository`. The
goal is for `RepositoryExt` to to only have provided methods, i.e. only
trivial (or empty) impls.

We previously thought of having `RepositoryExt` require all those new
traits. However, some of them will only actually used in a few or a
single function. The trait `Traversible` that we recently introduced
belongs into that category: it is only used in `issue_with_message` (and
fns in `Issue`, but that's a separate matter). We thus decided to
require `Traversible` only for that fn rather than the entire trait.