reference: add mod for reference related utilities
Currently, `libgitdit` is based on `git2`.  We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions.

References are quite essential for git-dit, as some vital information is
encoded in them. Previously, all that was spread over `issue` and other
modules. However, when decoupling all that from `git2` we'd rather not
sprinkle utilities all over the place but have clean interfaces for
interacting with references.

For these new interfaces we introduce a new module.