reference: add Reference trait
Currently, `libgitdit` is based on `git2`.  We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions.

References are quite essential for git-dit, as some vital information is
encoded in them. Previously, all that was spread over `issue` and other
modules. This change now introduces a trait with basic querying fns that
we implement for `git2::Reference` and future alternatives.