reference: add Store trait
Currently, `libgitdit` is based on `git2`. We extend `git2::Prepository`
via our trait `RepositoryExt`. We do want to support using other git
implementations in the future, namely `gix` and are thus currently
introducing abstractions which allow removal of hard-coded `git2` types
from the trait.

We decided to take the opportunity to modularize a bit more and move
some functionality into different traits that are then implemented for
`git2::Repository`. The goal is for `RepositoryExt` to require those,
and for it to only have provided methods, i.e. only trivial (or empty)
impls.

We already introduced a trait for references. However, we lacked a git
implementation agnostic method for retrieving or updating those
references. This change introduces a trait for doing so. It is currently
empty, and will be extended in the future with fns for retrieving and
setting references.