repository: require Store for RespositoryExt
Currently, `libgitdit` is based on `git2`. We extend `git2::Prepository`
via our trait `RepositoryExt`. We do want to support using other git
implementations in the future, namely `gix` and are thus currently
introducing abstractions which allow removal of hard-coded `git2` types
from the trait.

We decided to take the opportunity to modularize a bit more and move
some functionality into different traits that are then implemented for
`git2::Repository`. The goal is for `RepositoryExt` to require those,
and for it to only have provided methods, i.e. only trivial (or empty)
impls.

This change makes `RepositoryExt` require `reference::Store`, a trait we
recently introduced. In the future, this trait will allow us to retrieve
and update references, and thus issues. While we chose to require
`traversal::Traversible` only for individual fns, references are too
central for issue management and are used in most of `RepositoryExt`'s
fns.