reference: move associated type Base::Reference to Store
We recently introduced `reference::Store` as a future abstraction for
retrieving and updating references. Introducing that trait thus created
a proper place for a `Reference` associated type to life.

We previously placed it in `Base` since we made `error::Error` require
it, sicne we originally placed the conversion of a reference to a name
suitable for inclusion `'static` `Error`s in the `error::InnerError`
trait. However, we now have `Reference::name`. Thus, we desided to
simply require for `Reference::Name` to be convertible to
`error::IntoError::RefName`. This obsoletes the conversion utility in
`error::InnerError` and removes requirements on it. It will be removed
in the future.