reference: add an enum for representing kinds of references
Currently, `libgitdit` is based on `git2`.  We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions.

References are quite essential for git-dit, as some vital information is
encoded in them. Previously, all that was spread over `issue` and other
modules. However, when decoupling all that from `git2` we'd rather not
sprinkle utilities all over the place but have clean interfaces for
interacting with references.

We sometimes need to figure out whether a reference is a head or leaf
reference. We do have `issue::IssueRefType` for representing that
information, but the `Any` is a bit awkward when querying. This change
now introduces a new enum that also holds information about the leaf
`Oid`. In the future, we will introduce a fn which allows querying the
`Kind` of a `Reference`.