reference: add fn for extracting parts of a reference
Currently, `libgitdit` is based on `git2`.  We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions.

References are quite essential for git-dit, as some vital information is
encoded in them. We may want to sometimes extract that information from
a given reference name. We do already have `issue::IssueRefType::of_ref`
for this purose, but we'd rather have a fn associated to the `Reference`
trait.