reference: add Reference impl for testing
Currently, `libgitdit` is based on `git2`.  We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions. To this end, we recently introduced
the `Base` trait.

As a side effect, those traits also allow testing some more generic
parts of the library, i.e. provided associated and generic fns. For
those tests, we need mock types that we can substitute via those
abstractions. This change introduces a mock for `Reference`s.