base: impl PartialEq<&str> for TestOid
We recently introduced the `TestOid` type for git impl agnostic tests.
Those tests will, in many cases, contain a comparison between an id
resulting from some sort of computation and an expected value. Being
able to directly compare a `TestOid` to a `str` makes formulating this a
bit more pleasant.