remote: add trait for abstracting remotes' names
Currently, `libgitdit` is based on `git2`.  We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions.

References are quite essential for git-dit, as some vital information is
encoded in them. We need to construct references and reference parts for
querying, without relying on `git2`-specific functionality. This
obviously also involves remotes. We thus need an abstraction for
handling the implementation specific remote names and allow constructing
the assocaited reference parts.