issue: make Issue generic over the repository impl
Currently, `libgitdit` is based on `git2`.  We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions.

One of the abstractions was the `Base` trait, which provides some basic
associated types. That trait, as well as other traits requireing it,
will be implemented for applicable types representing an repository (or
some aspect of one).

`Issue`, a very central and prominent type, is still a heavy user of
`git2::Repository` and its fns. This change makes at least the data type
itself and the id querying fn generic over all reposity types (that impl
`Base`).