issue: add fn for creating an issue without doing any checks
`Issue`s are created through fns associated to `RepositoryExt`. Those
fns currently use `Issue::new`, which used to do a lookup of the issue's
initial message, an operation that might fail. However, we recently made
`Issue` carry only an id rather than an entire object. And `Issue::new`
is public, with little to no hint that it is supposed to be used with
validation.

This change introduces a new fn which is public only crate-wide, with a
name and documentation making it clear that the caller is responsible
for doing the necessary checking. It will replace `Issue::new`, which
will be removed in the near future.