repository: migrate RepositoryExt fns away from Issue::new
`Issue`s are created through fns associated to `RepositoryExt`. Those
fns previously used `Issue::new`, which might fail, is public and has
little to no hint that it is supposed to be used with validation.

We recently introduces `Issue::new_unchecked` as a replacement. This
change now migrates all users in `RepositoryExt` over to that new fn.