issue: remove now unused Issue::new
`Issue`s are created through fns associated to `RepositoryExt`. Those
fns previously used `Issue::new`, which had little to no hint that it
was supposed to be used with validation. We do not want a public API
like that. In fact, we intend for `Issues` to be created _exclusively_
via `RepositoryExt`. Thus, we remove this fn.