remote: make RemoteExt::issue_refspec use Issue::id rather than ref_part
`Issue::ref_part` does, by specification, nothing but convert the
issue's id into a `String`. Given that `Base::Oid` requires `Display`,
this is utterly pointless and just a waste of resources. We thus want to
get rid of it.

This change removes the only use outside the `issue` mod,
`RemoveExt::issue_refspec`.