remote: make Name::as_str and ref_path return a Result
We previously hid any errors generated in fns associated to some traits,
including `Name`. We want to change that and communicate those inner
errors. This change does so, and switches `Names::NameIter` for the
`git2::string_array::StringArray` impl to an `Iterator` that lets us do
the failible UTF-8 conversion ourselves.