reference: add fn for retrieving all remote reference paths
In the future, we will need to gather references from all remotes. This,
of course, involves retrieving the referrence base path for each remote.
We already have `remote::Names::ref_paths`. This change introduces a
convenience fn usable directly from `Store`.

Sadly, we need to collect the paths in the new `remote_ref_paths` fn,
sicne the iterator over those paths may have the lifetime of the
`remote::Names` through which it was created.