error: allow Error to carry a fmt::Error as inner error
In the future, we'll construct reference paths and potentially other
things as `String`s. We will likely do so using the `fmt::write!` macro,
which produces code that returns a `fmt::Result`. We don't want to
handle those `fmt::Error`s silently but convey them to users. Thus,
we'll need to wrap them in `error::Error` at some point.