error: add fn for changing the Kind of an Error
Currently, `libgitdit` is based on `git2`.  We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions. Some of those abstractions return
`error::Result`.

In some cases, we may want to switch out the `Kind` with something more
specific than we could provide from under the abstraction. This change
introduces a fn that lets us do that.