iter: remove MessageExt trait
We plan to remove the entirety of the `message` module, since it will
serve little purpose in the future. As a preparation, we remove the
`MessageExt` trait that used to provide some utilities based on the
`Message` trait. Its only purpose was providing a fn for the
accumulation of trailers. If we decide to pursue an information model
making heavy use of trailers in the future, we'll introduce new, better
[TM] utilities for that.