message: remove module
In the past, the `message` module, and the `Message` trait in
particular, served two purposes: decomposing a commit message for the
extraction of metadata (e.g. subject, trailers) and providing general
text processing utilities (e.g. quoting, trimming). Both kinds of
functions would be implemented on text, under heavy use of custom
`Iterators` (return position `impl` wasn't a thing when we wrote those).

As for the extraction part, `git2::Commit` does already provide
utilities for extracting parts of a commit messages and if/when we start
supporting other git implementations we'd rather try to have such fns
upstream, sicne they are certainly not only useful to us. Naturally,
we'd still want an intermediate trait for the sake of implementation
agnostic code, but that should not be hung up on representing a message
consisting of blocks.

And we need to extract trailers only if we decide that we need to
accumulate trailers to decide the state of an issue in the future in the
first place.

As for general text processing utilities: most, if not all of them, were
trivial and had little to nothing to do with actual issue tracking.
Hence, that functionality is best provided from inside any user, or as a
separate crate (that probably already exists).

Therefore, we decided that the `message` module is dead weight, and that
we don't want to put any more effort into it. Leaving it as is will
probably hinder us during further developent. So out it goes.