reference: do not require exessive lifetime for Reference::target
We went over board with the lifetime hacks for `Referenbce`. They start
to hurt and obstruct rather than faciliate git implementation agnostic
code.

In the case of `Reference::target`, we are absolutely sure that we never
ever want the returned `Oid` to depend on `'r`. Thus, there is no need
to specify `'r` for `self`. In fact, we do not want to require any
lifetime for `self`.