reference: make Reference::name return the name by ref rather than value
We went over board with the lifetime hacks for `Referenbce`. They start
to hurt and obstruct rather than faciliate git implementation agnostic
code.

In the case of `Reference::name`, we originally needed to accomodate
some ill-fated lossy conversion in the `git2` impl. That is gone, as we
decided to propagate those errors. Now, we can just state that whatever
the type of the name is, its returned by reference.

We thus can also remove the `'r` lifetime from `self` for the affected
fns. As that lifetime is also a requirement, it will ultimately create
issues.