reference: remove lifetime parameter from Reference trait
We went over board with the lifetime hacks for `Referenbce`. They start
to hurt and obstruct rather than faciliate git implementation agnostic
code.

We originally needed to accomodate some ill-fated lossy conversion in
the `git2` impl. Those are now gone, as we decided to propagate those
errors, and with them the need for a lifetime parameter for the trait.