object: add new module for object related facilities
Currently, `libgitdit` is based on `git2`. We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions.

We don't yet have the necessary git agnostic abstractions for retrieving
and creating commits or trees. This change introduces a new module that
will house those facilties.