object: add trait for object databases
Currently, `libgitdit` is based on `git2`. We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions.

We did not yet have the necessary git agnostic abstractions for
retrieving and creating commits or trees. This change introduces a new
trait that allows querying commits and trees.