object: add trait for commits in dedicated submodule
Currently, `libgitdit` is based on `git2`. We do want to support using
other git implementations in the future, namely `gix`, and are thus
currently introducing abstractions.

We currently lack a git agnostic abstraction for one of the most central
entities in git: the commit. This change introduces a submodule for that
abstraction, which we will introduce in the future.