issue: make Issue::initial_message generic
Currently, `libgitdit` is based on `git2`.  We do want to support using
other git implementations in the future, namely `gix`. To this end, we
recently introduced various abstractions. This includes the `Database`
trait, which allows retrieval of objects by their `Oid`. And this is all
we need to implement `Issue::initial_message`.