error: impl From<{I, Utf8Error, fmt::Error}> for Inner
We want to impl `ResultExt` for all potential inner error types, not
just the `InnerError` used by the underlying git implementation. This
involves formulating a trait bound based on the `Inner` type. What we're
interested in is the capacity to convert an error into an `Inner`. And
that requires some means of conversion other than explicit construction
via the variants.

So we do what we should have done when we introduced `Inner` and impl
`From` all the wrapped error types for it.