error: make Error::with_inner take any potential inner error
We want to impl `ResultExt` for all potential inner error types, not
just the `InnerError` used by the underlying git implementation. In
particular, we want to impl `ResultExt` for all `Into<Inner>`. Since we
use `Error::with_inner` in the `ResultExt` impl, we need to adapt it.

Since that fn now uses `Inner` in its signature, it needs to be public
only within the crate. Since we only use it within the `error` module
anyway, we chose to make it private.