error: make Inner public within the crate
We want to impl `ResultExt` for all potential inner error types, not
just the `InnerError` used by the underlying git implementation. This
involves formulating a trait bound based on the `Inner` type.

Since `ResultExt` is public within the crate, we thus need `Inner` to be
public within the crate since it will be leaked through the trait bound.