error: add fn for convenient wrapping of a Kind
We recently started allowing to wrap any potential inner error in an
`Error` via the `ResultExt` trait. However, we sometimes we also want to
operate on the errors themselves.

We do do have various `From` impls for `Error`, but those rely on
`Kind::Other`, and we want to get rid of that. In other words, we always
need to supply a `Kind` explicitly. Thus, we want to convert `Kind` to
`Error` first and then add the (optional) inner error.

However, we chose to make `Error::with_inner` private. We could make it
public within the crate, but instead we chose to introduce a more
convenient fn for constructing an `Error` based on a `Kind` taking an
inner error.