base: impl AddAssign<u8> and Default for TestOid
We introduced `TestOid` in order to formulate tests. For those tests, we
need to construct those. They do not need to reflect any object's
contents, but we do want newly generated ids to be unique per test. The
easiest way to achieve this is to compute them by simple incrementation.