base: impl Ord for TestOid
We plan to introduce a mock `TraversalBuilder` in the future. This will
need to create an `Iterator` yielding commits in topological order.
Because we don't want to get insane, we do _not_ want to replicate all
the logic a real implementation needs to have to achieve this.

Now, we already decided that we'll compute `TestOid`s by simple
incrementation. If we don't mess this up, we can order any commits
topologically by simply ordering by theit ids.