object: add R lifetime to commit::FollowUp
`FollowUp::follow_up` receives a `Base` which the receiver may use to
perform some arbitrary action such as updating a reference. That only
works because `database` is not only a `Base` but a `Database` and
often (usually) also a `reference::Store`. And many of those fns take
`&self` with an explicit lifetime. If we don't want to run into lifetime
related troubles, we do ned to forward that to users in `follow_up`.