issue: migrate tests to dedicated testing impl
Everything but impls for foundational traits should be agnostic wrt. the
underlying git implementation. This includes tests, and we did introduce
impls dedicated for testing that also does not need any access to the
file system. Now that we made `Issue` completely agnostic, we can
migrate all tests in this module away from `git2`.