repository: migrate tests to dedicated test repo
Everything but impls for foundational traits should be agnostic wrt. the
underlying git implementation. This includes tests, and we did introduce
impls dedicated for testing that also does not need any access to the
file system. Now that we made `RepositoryExt` mostly agnostic, we can
migrate the tests in this module away from `git2`.