repository: remove RepositoryExt::create_issue
We want `RepositoryExt` to only have provided fns, and the only fn that
isn't is `RepositoryExt::create_issue`. We could rewrite it to make it a
provided fn, but we recently introduced `RepositoryExt::issue_builder`
which already covers the use-case while being more ergonomic. And unlike
with `Issue::add_message` we don't have to accomodate tests or dependent
code during the transition. We therefore decided to just drop this fn.