gc: make CollectableRefs::for_issue generic over all repo types
Currently, `libgitdit` is based on `git2`.  We do want to support using
other git implementations in the future, namely `gix`. To this end, we
recently introduced various abstractions.

These changes make use of those abstractions and make the most central
fn of `CollectableRefs` generic over all repository types. This also
causes `CollectableRefs::repo` to not be used any more. We will remove
that in the member in the future.