gc: redefine how CollectableRefs are created
Previously, `CollectableRefs` would carry a repository reference that it
used for various purposes. Which, in hindsight, was completely pointless
given that `Issue` already carries a reference to the repository it
originated from. This change removes that unused member, together with
`CollectableRefs::new`. From now on, we will not need a repository for
construction of a `CollectableRefs`. Instead we can just use `Default`,
or copy an existing `CollectableRefs`.