repository: rename RepositoryExt to Repository
We origninally added the `Ext` to avoid conflicts with
`git2::Repository`. But things changed since then. We are much more
comfortable to use types from other modules and crates directly. And in
the case of `git2` we specifically decided to avoid imports in order to
facilitate future feature-gating of the `git2` dependency.