trailer: rewrite FromStr impl for Trailer with only std
Parsing `Trailer`s is easy, but back when we wrote this we, for some
reason, choose to use a regex, and thus the `Regex` crate. And because
the `Regex` needed to be parsed, we also pulled in `lazy_static`.

This was clearly overkill, especially considering that we don't use
those external dependencies anywhere else. Rewriting this allows us to
get rid of those deps.

Since we don't want a hard dependency on any specific `InnerError`, we
decided to just return the input in place of an error. The `error::Kind`
we would have chosen wouldn't be any more specific anyway.