remote: use full path for import of Issue
We are in the process of modernizing the code to make it easier to work
on it in the future. This includes upgrading to a newer edition (2024),
or rather, specifying an edition in the first place. Currently, we
default to the edition 2015, which is technically correct.

In a few instances, we still use legacy imports for internal modules.
This change cleans up one import in the `remote` module.