repository: use full paths for toplevel imports
We are in the process of modernizing the code to make it easier to work
on it in the future. This includes upgrading to a newer edition (2024),
or rather, specifying an edition in the first place. Currently, we
default to the edition 2015, which is technically correct.

In a few instances, we still use legacy imports for internal modules.
This change cleans up imports in the `repository` module. This does
result in "unused imports" for now, but they will be used when legacy
imports change semantics as we switch to the new edition.