Switch to Rust edition 2024
Back when we started this projects, Rust editions did not yet exist. We
thus did not specify any initially and never came around to do so.
Because of that, we defaulted to edition 2015. In that edition, imports
and "deconstruction" of enum variants (in match statements) work very
differently than we are used from modern Rust.

But thanks to recent efforts, we now can switch to edition 2024 with
only a few changes being neccessary.